simulation/python/requirements.txt
absl-py==0.14.1
astor==0.8.1
cached-property==1.5.2
control-toolbox==0.1.0
cycler==0.10.0
gast==0.5.2
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.4.0
importlib-metadata==4.8.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mpmath==1.2.1
numpy==1.21.2
pandas==1.3.4
Pillow==8.4.0
protobuf==3.18.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.1
six==1.16.0
sympy==1.9
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
wrapt==1.13.2
zipp==3.6.0
